♻️ refactor: remove GiftCardReducer, replace useReducer with useState + createGiftCard from core
- Moved GiftCardI, GiftCardRecipientI, GiftCardState, giftCardInitialState into GiftCardContext.ts
- Replaced useReducer in GiftCard.tsx with useState (errors, loading, giftCardRecipient)
- Replaced useReducer in GiftCardContainer.tsx with useState
- Inlined addGiftCard business logic using createGiftCard from @commercelayer/core
- Deleted GiftCardReducer.ts
